+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [Unreleased]
 
 Rails support!
-Need a better way to find Rails config, going to release 0.1.0 after that. 
+Need a better way to find Rails config, going to release 0.1.0 after that.
+Going to change the name of the gem to something a bit more original than HardWorker. 
+(HardWorker is already in use in Sidekiq documentation) 
 ## [0.0.1] - 2021-07-15
 
 - Initial release! :tada:
@@ -16,3 +18,9 @@ Things are almost working. You can enqueue jobs! That should be more than enough
 
 - YAML Marshalled queue! So now the queue stays in memory even if restarted. Though kill and term signals still need catching.. that is not done yet. 
 - Also, now you can use classes! As long as the class has a perform-method. 
+
+## [0.0.4] - 2021-07-24
+
+- Rails support!
+- Banner (for Belated though...)
+- Final release, as the name of this gem is going to change to Belated.